+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog of threedigrid-builder
 1.0.3 (unreleased)
 ------------------
 
+- Write "grid_coordinate_attributes" also for pure 1D models.
+
 - Make requesting spatial reference of GDAL dataset compatible with GDAL 2.x.